Move CensusContext class too
